% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimond_plot.R
\name{diamond_plot}
\alias{diamond_plot}
\title{Title}
\usage{
diamond_plot(df, varx, vary)
}
\arguments{
\item{df}{a dataframe}

\item{varx}{a string name for variable x in the dataframe df}

\item{vary}{a string name for variable y in the dataframe df}
}
\value{
a plot of dimond makers of \code{vary} verus \code{varx}
}
\description{
This fumction produces a ggplot plot of \code{vary} versus \code{varx}
with diamond markers
}
\examples{
df.test <- data.frame(x = rnorm(10),
                      y = rnorm(10))
diamond_plot(df.test, 'x', 'y')

}
